requirements.txt
aiohttp==3.8.3
python-telegram-bot==20.0
pyyaml==6.0
aiohttp-jinja2==1.5.1
aiosqlite==0.18.0